ERP/DAP/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
requests==2.32.3
python-dotenv==1.0.1
openpyxl==3.1.5
anthropic>=0.40.0
prophet>=1.1.5
pandas>=2.0.0
